HSEARCH-1375 Introduce FailureContext, a more generic version of IndexFailureContext
